Refactor: separate transaction builders from lib.rs
